refactor(config): move `request`-specific config to backends/request/
This is in preparation to deprecate the `request`-based backend.
