+++ lib/backends/request/client.test.js
@@ -4,7 +4,7 @@
 const { expect } = require('chai')
 const nock = require('nock')
 
-const Request = require('./request')
+const Request = require('./client')
 
 const url = 'http://mock.kube.api'
