+++ lib/backends/request/index.js
@@ -1,0 +1,5 @@
+const Request = require('./client')
+
+Request.config = require('./config')
+
+module.exports = Request
