+++ lib/index.js
@@ -2,5 +2,5 @@ module.exports = {
   Client: require('./swagger-client').Client,
   Client1_13: require('./swagger-client').Client1_13,
   alias: require('./alias'),
-  config: require('./config')
+  config: require('./backends/request/config')
 }
